src/main/java/cl/uchile/dcc/scrabble/models/operation/arithmetic/Div.java
package cl.uchile.dcc.scrabble.models.operation.arithmetic;

import cl.uchile.dcc.scrabble.models.operation.OperableEntity;
import cl.uchile.dcc.scrabble.models.operation.Operation;
import cl.uchile.dcc.scrabble.models.operation.constant.Constant;

public class Div implements Operation {
    private OperableEntity leftEntity;
    private OperableEntity rightEntity;

    public Div(OperableEntity leftOperation, OperableEntity rightOperation){
        this.leftEntity = leftOperation;
        this.rightEntity = rightOperation;
    }

    @Override
    public Constant evaluate() {
        return this.leftEntity.evaluate().div(
                this.rightEntity.evaluate()
        );
    }
}
